Remove "__global__" stream name, use "all".
In previous versions we had a kind of global/default stream called
"all". You could not have an event in more than one stream so it landed
either in "all" stream or the one you specified on publish.

This is rather confusing -- stream "all" that does not have all events.

Then was the "__global__" stream introduced, which if fact had good
semantics for being called "all".

It is now possible to have all events always in one global/"all" stream.
This would actually fix "all" stream semantics -- let's stick with this
name instead of introducing new one.